web/feat: embedded pages cleanup & open in new tab
