+++ packages/web/app/[locale]/()/(pages)/projects/[id]/page.tsx
@@ -1,13 +1,17 @@
 import { HydrationBoundary, dehydrate } from '@tanstack/react-query';
-import { getQueryClient } from '@/domain/queries/server-query-client';
-import { ProjectPage as ClientProjectPage } from '@/app/_components/project/project-page';
+import { TaskStatus } from '@prisma/client';
 import { withPrivatePage } from '@/app/_hoc/with-private-page';
+import { getQueryClient } from '@/domain/queries/server-query-client';
 import { notesQuery } from '@/domain/queries/notes-query';
 import { tasksQuery } from '@/domain/queries/tasks-query';
-import { TaskStatus } from '@prisma/client';
 import { quickLinksQuery } from '@/domain/queries/quick-links-query';
 import { foldersQuery } from '@/domain/queries/folders-query';
 import { folderQuery } from '@/domain/queries/folder-query';
+import { projectQuery } from '@/domain/queries/project-query';
+import { embeddedPagesQuery } from '@/domain/queries/embedded-pages-query';
+import { ProjectPage as ClientProjectPage } from '@/app/_components/project/project-page';
+import { PageContainer } from '@/app/_components/page-container';
+import { NotificationsWidget } from '@/app/_components/notifications-widget';
 
 type Params = { locale: string; id: string };
 type Props = {
@@ -20,7 +24,10 @@ async function ProjectPage(props: Props) {
 
   return (
     <HydrationBoundary state={dehydratedState}>
-      <ClientProjectPage id={id} />
+      <PageContainer>
+        <ClientProjectPage id={id} />
+      </PageContainer>
+      <NotificationsWidget projectId={id} />
     </HydrationBoundary>
   );
 }
@@ -29,6 +36,10 @@ async function useProjectPage({ params: { id }, searchParams }: Props) {
   const folderParamsId = searchParams.folder ?? 'null';
   const queryClient = getQueryClient();
   await Promise.all([
+    queryClient.prefetchQuery({ queryKey: projectQuery.key(id) }),
+    queryClient.prefetchQuery({
+      queryKey: embeddedPagesQuery.key(id),
+    }),
     queryClient.prefetchQuery({ queryKey: notesQuery.key(id) }),
     queryClient.prefetchQuery({
       queryKey: tasksQuery.key(id, {
