+++ packages/web/navigation/paths.ts
@@ -4,8 +4,6 @@ export const paths = {
   register: () => '/register',
   projects: () => '/projects',
   project: (id: string) => `/projects/${id}`,
-  embeddedPage: (projectId: string, pageId: string) =>
-    `/projects/${projectId}/embedded-pages/${pageId}`,
   todo: () => '/to-do',
   calendar: () => '/calendar',
 };
